fixed some bug

+++ src/toast/ops/kernels/jax_kernels/math/healpix.py
@@ -7,7 +7,7 @@ import numpy as np
 import jax.numpy as jnp
 
 TWOINVPI = 0.63661977236758134308
-MACHINE_EPSILON = np.finfo(np.float).eps
+MACHINE_EPSILON = np.finfo(np.cfloat).eps
 
 
 class HPIX_JAX:
